requirements.txt
opencv-python
pillow
imagehash
exifread
dlib
insightface
qdrant-client
chromadb
streamlit
gradio
crewai
# Event/Milestone extraction
torch
torchvision
torchaudio
git+https://github.com/openai/CLIP.git